requirements.txt
tensorflow~=2.9.1
tensorflow_addons
focal_loss
segmentation_models
PyYAML
numpy
scikit-learn
PyYAML
tqdm
matplotlib
scipy
plyfile~=0.7.4